+++ CHANGELOG.md
@@ -1,9 +1,21 @@
 # Change Log
 
-All notable changes to the "yai" extension will be documented in this file.
+## 0.0.1
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+### Golang
 
-## [Unreleased]
+* Support for Indexing go.mod, standard libs and local packages imports
+* Support for importing modules with quick pick and quick input
+* Ranked aliases of go modules by count of usage
 
-- Initial release
+## 0.0.2
+
+### Golang
+
+* Added configuration group "Yet Another Importer" for excluding paths from indexing
+* Parsed go.mod to get module info and indexed all local sub packages in the workspace
+
+### Others
+
+* Set a extension icon :)
+* Added nyc dependency for code coverage